Secure all entry points

+++ routers/test/test_admin.py
@@ -2,10 +2,13 @@ from fastapi import APIRouter, Body, Path, Depends
 from sqlalchemy.orm import Session
 from starlette import status
 
+from controller.login import get_current_token, check_type_user
 from db.database import get_db
 from db.orm import admins_orm
+from db.orm.exceptions_orm import credentials_exception
 from schemas.admin_base import AdminDisplay, AdminRequest
 from schemas.basic_response import BasicResponse
+from schemas.token_base import TokenSummary
 
 router = APIRouter(
     prefix='/test/admin',
@@ -20,8 +23,12 @@ router = APIRouter(
 )
 async def create_admin(
         request: AdminRequest = Body(...),
-        db: Session = Depends(get_db)
+        db: Session = Depends(get_db),
+        current_token: TokenSummary = Depends(get_current_token)
 ):
+    if not check_type_user(current_token, is_a='admin'):
+        raise credentials_exception
+
     response = admins_orm.create_admin(db, request)
 
     return response
@@ -34,8 +41,12 @@ async def create_admin(
 )
 async def get_admin(
         id_admin: str = Path(..., min_length=12, max_length=49),
-        db: Session = Depends(get_db)
+        db: Session = Depends(get_db),
+        current_token: TokenSummary = Depends(get_current_token)
 ):
+    if not check_type_user(current_token, is_a='admin'):
+        raise credentials_exception
+
     response = admins_orm.get_admin_by_id_admin(db, id_admin)
 
     return response
@@ -48,8 +59,12 @@ async def get_admin(
 )
 async def get_admin_by_id_user(
         id_user: int = Path(..., gt=0),
-        db: Session = Depends(get_db)
+        db: Session = Depends(get_db),
+        current_token: TokenSummary = Depends(get_current_token)
 ):
+    if not check_type_user(current_token, is_a='admin'):
+        raise credentials_exception
+
     response = admins_orm.get_admin_by_id_user(db, id_user)
 
     return response
@@ -63,8 +78,12 @@ async def get_admin_by_id_user(
 async def update_admin(
         id_admin: str = Path(..., min_length=12, max_length=49),
         request: AdminRequest = Body(...),
-        db: Session = Depends(get_db)
+        db: Session = Depends(get_db),
+        current_token: TokenSummary = Depends(get_current_token)
 ):
+    if not check_type_user(current_token, is_a='admin'):
+        raise credentials_exception
+
     response = admins_orm.update_admin(db, request, id_admin)
 
     return response
@@ -77,8 +96,12 @@ async def update_admin(
 )
 async def delete_admin(
         id_admin: str = Path(..., min_length=12, max_length=49),
-        db: Session = Depends(get_db)
+        db: Session = Depends(get_db),
+        current_token: TokenSummary = Depends(get_current_token)
 ):
+    if not check_type_user(current_token, is_a='admin'):
+        raise credentials_exception
+
     response = admins_orm.delete_admin(db, id_admin)
 
     return response
